Feat: New toxicity metric & migrate to uv
