+++ .gitignore
@@ -1,10 +1,49 @@
 .DS_Store
+
+# Virtual environments
 venv/
+.venv/
+env/
+ENV/
+
+# Build artifacts
 dist/
+build/
+*.egg-info/
 alquimia_fair_forge.egg-info/
+
+# Python cache
 __pycache__/
+*.pyc
+*.pyo
+*.pyd
+.Python
+
+# Distribution / packaging
 *.gz
+*.whl
+
+# Logs
 *.log
+
+# LaTeX
 *.aux
 *.out
+
+# Jupyter
 .ipynb_checkpoints
+*.ipynb_checkpoints/
+
+# uv
+uv.lock
+
+# IDEs
+.vscode/
+.idea/
+*.swp
+*.swo
+*~
+
+# OS
+.DS_Store
+Thumbs.db
