+++ fair_forge/metrics/__init__.py
@@ -1,5 +1,6 @@
 from .humanity import Humanity
 from .bias import Bias
+from .toxicity import Toxicity, HurtlexLoader
 from .context import Context
 from .conversational import Conversational
 from .agentic import Agentic
@@ -8,6 +9,8 @@ from .bestOf import BestOf
 __all__ = [
     'Humanity',
     'Bias',
+    'Toxicity',
+    'HurtlexLoader',
     'Context',
     'Conversational',
     'Agentic',
