+++ fair_forge/metrics/bias.py
@@ -1,49 +1,23 @@
 from pydantic import BaseModel
-from fair_forge import FairForge, Retriever, Guardian,ToxicityLoader
-from typing import Optional, Type
-from fair_forge.schemas import Batch, ProtectedAttribute, BiasMetric, ToxicityDataset
+from fair_forge import FairForge, Retriever, Guardian
+from typing import Type
+from fair_forge.schemas import Batch, ProtectedAttribute, BiasMetric
 import scipy.stats as st
-from sentence_transformers import SentenceTransformer
-from hdbscan import HDBSCAN
-import pkg_resources
-import pandas as pd
-from collections import Counter
-from umap import UMAP
 from tqdm import tqdm
-import numpy as np
-
-class HurtlexLoader(ToxicityLoader):
-    def load(self,language:str) -> list[ToxicityDataset]:
-        df = pd.read_csv(pkg_resources.resource_filename(
-            "fair_forge", f"artifacts/toxicity/hurtlex_{language}.tsv"
-        ), sep="\t", header=0)
-        return [ToxicityDataset(word=row["lemma"], category=row["category"]) for _, row in df.iterrows()]
 
 
 class Bias(FairForge):
     """
     A class for measuring and analyzing bias in AI assistant responses.
-    
+
     This class implements various methods to detect and quantify bias across different protected attributes
     such as gender, race, religion, nationality, and sexual orientation. It uses a combination of
-    clustering techniques, confidence intervals, and guardian-based bias detection to provide
-    comprehensive bias analysis.
+    confidence intervals and guardian-based bias detection to provide comprehensive bias analysis.
 
     Attributes:
         protected_attributes (list[ProtectedAttribute]): List of protected attributes to monitor for bias
         guardian (Guardian): Instance of the Guardian class for bias detection
         confidence_level (float): Confidence level for statistical calculations (default: 0.95)
-        embedding_model (SentenceTransformer): Model for generating text embeddings
-        toxicity_loader (ToxicityLoader): Loader for toxicity-related data
-        min_cluster_size (int): Minimum size for clusters in HDBSCAN
-        cluster_selection_epsilon (float): Epsilon value for cluster selection
-        cluster_selection_method (str): Method used for cluster selection
-        umap_n_components (int): Number of components for UMAP dimensionality reduction
-        umap_n_neighbors (int): Number of neighbors for UMAP
-        umap_min_dist (float): Minimum distance parameter for UMAP
-        umap_random_state (int): Random state for UMAP reproducibility
-        umap_metric (str): Metric used for UMAP
-        toxicity_cluster_use_latent_space (bool): Whether to use latent space for clustering
     """
 
     class ClopperPearson(BaseModel):
@@ -70,17 +44,6 @@ class Bias(FairForge):
         retriever: Type[Retriever],
         guardian: Type[Guardian],
         confidence_level: float = 0.95,
-        embedding_model:str = "all-MiniLM-L6-v2",
-        toxicity_loader: Type[ToxicityLoader] = HurtlexLoader,
-        toxicity_min_cluster_size:int = 5,
-        toxicity_cluster_selection_epsilon:float = 0.01,
-        toxicity_cluster_selection_method:str = "euclidean",
-        toxicity_cluster_use_latent_space:bool = True,
-        umap_n_components:int = 2,
-        umap_n_neighbors:int = 15,
-        umap_min_dist:float = 0.1,
-        umap_random_state:int = 42,
-        umap_metric:str = "cosine",
         **kwargs,
     ):
         """
@@ -90,17 +53,6 @@ class Bias(FairForge):
             retriever (Type[Retriever]): The retriever class to use
             guardian (Type[Guardian]): The guardian class for bias detection
             confidence_level (float, optional): Confidence level for statistical calculations. Defaults to 0.95.
-            embedding_model (str, optional): Name of the embedding model to use. Defaults to "all-MiniLM-L6-v2".
-            toxicity_loader (Type[ToxicityLoader], optional): Loader for toxicity data. Defaults to HurtlexLoader.
-            toxicity_min_cluster_size (int, optional): Minimum cluster size for HDBSCAN. Defaults to 5.
-            toxicity_cluster_selection_epsilon (float, optional): Epsilon for cluster selection. Defaults to 0.01.
-            toxicity_cluster_selection_method (str, optional): Method for cluster selection. Defaults to "euclidean".
-            toxicity_cluster_use_latent_space (bool, optional): Whether to use latent space. Defaults to True.
-            umap_n_components (int, optional): Number of UMAP components. Defaults to 2.
-            umap_n_neighbors (int, optional): Number of UMAP neighbors. Defaults to 15.
-            umap_min_dist (float, optional): Minimum distance for UMAP. Defaults to 0.1.
-            umap_random_state (int, optional): Random state for UMAP. Defaults to 42.
-            umap_metric (str, optional): Metric for UMAP. Defaults to "cosine".
             **kwargs: Additional keyword arguments passed to the parent class
         """
         super().__init__(retriever, **kwargs)
@@ -125,38 +77,19 @@ class Bias(FairForge):
                 attribute=ProtectedAttribute.Attribute.sexual_orientation,
                 description="Sexual orientation and romantic attraction, including but not limited to heterosexual, homosexual, bisexual, pansexual, asexual, and other orientations. This attribute is vital for detecting LGBTQ+ discrimination and ensuring equal treatment regardless of sexual orientation."
             )
-        ] ## PROTECTED ATTRIBUTES DEFINED BY FAIR FORGE
+        ]  # PROTECTED ATTRIBUTES DEFINED BY FAIR FORGE
 
         self.guardian = guardian(**kwargs)
         self.confidence_level = confidence_level
-        self.embedding_model = SentenceTransformer(embedding_model)
-        self.toxicity_loader = toxicity_loader()
-        self.min_cluster_size = toxicity_min_cluster_size
-        self.cluster_selection_epsilon = toxicity_cluster_selection_epsilon
-        self.cluster_selection_method = toxicity_cluster_selection_method
-        self.umap_n_components = umap_n_components
-        self.umap_n_neighbors = umap_n_neighbors
-        self.umap_min_dist = umap_min_dist
-        self.umap_random_state = umap_random_state
-        self.umap_metric = umap_metric
-        self.toxicity_cluster_use_latent_space = toxicity_cluster_use_latent_space
 
-        self.logger.info(f"--CONFIGURATION--")
-        self.logger.debug(f"Min cluster size: {self.min_cluster_size}")
-        self.logger.debug(f"Cluster selection epsilon: {self.cluster_selection_epsilon}")
-        self.logger.debug(f"Cluster selection method: {self.cluster_selection_method}")
-        self.logger.debug(f"Cluster use latent space: {self.toxicity_cluster_use_latent_space}")
+        self.logger.info(f"--BIAS CONFIGURATION--")
         self.logger.debug(f"Confidence level: {self.confidence_level}")
-        self.logger.debug(f"UMAP n components: {self.umap_n_components}")
-        self.logger.debug(f"UMAP n neighbors: {self.umap_n_neighbors}")
-        self.logger.debug(f"UMAP random state: {self.umap_random_state}")
-        self.logger.debug(f"UMAP metric: {self.umap_metric}")
 
-        ## List protected attributes
+        # List protected attributes
         for attribute in self.protected_attributes:
             self.logger.debug(f"Protected attribute: {attribute.attribute.value}")
 
-    def _clopper_pearson_confidence_interval(self,samples:int,k_success:int)->ClopperPearson:
+    def _clopper_pearson_confidence_interval(self, samples: int, k_success: int) -> ClopperPearson:
         """
         Calculate the Clopper-Pearson confidence interval for a binomial proportion.
 
@@ -176,61 +109,8 @@ class Bias(FairForge):
             p_u = st.beta.ppf(1 - alpha/2, k_success + 1, samples - k_success)
 
         p_l = st.beta.ppf(alpha/2, k_success, samples - k_success + 1)
-        
+
         return self.ClopperPearson(lower_bound=p_l, upper_bound=p_u, probability=p_truth, samples=samples, k_success=k_success, alpha=alpha)
-    
-    def _cluster_profiling(self,batch:list[Batch],language:str) -> tuple[dict[float,float],list,list,list]:
-        """
-        Perform clustering analysis on assistant responses to identify potential bias patterns.
-
-        Args:
-            batch (list[Batch]): List of batch interactions to analyze
-            language (str): Language of the responses
-
-        Returns:
-            tuple: Contains:
-                - dict[float,float]: Toxicity scores for each cluster
-                - list: UMAP reduced embeddings
-                - list: Original embeddings
-                - list: Cluster labels
-        """
-        assistant_answers = [i.assistant for i in batch]
-        embeddings = self.embedding_model.encode(assistant_answers)
-
-        reducer = UMAP(n_components=self.umap_n_components, 
-                       random_state=self.umap_random_state, 
-                       n_neighbors=self.umap_n_neighbors, 
-                       metric=self.umap_metric, 
-                       min_dist=self.umap_min_dist
-        )
-        
-        clusterable_embeddings = reducer.fit_transform(embeddings)
-        
-        clusterer = HDBSCAN(
-            min_cluster_size=self.min_cluster_size, 
-            metric=self.cluster_selection_method, 
-            cluster_selection_epsilon=self.cluster_selection_epsilon,
-            prediction_data=True
-        )
-        
-        labels = clusterer.fit_predict(clusterable_embeddings if self.toxicity_cluster_use_latent_space else embeddings)
-        
-        toxicity_datasets = self.toxicity_loader.load(language)
-        score_cluster = {}
-
-        for lbl in set(labels):
-            # 1. Collect all the assistant answers that belong to the same cluster
-            texts = [resp for resp, l in zip(assistant_answers, labels) if l == lbl]
-
-            # 2. Count the number of times each word appears in the assistant answers & toxic words
-            cnt = Counter(word for t in texts for word in t.lower().split())
-            toxic_words = sum(cnt[t.word] for t in toxicity_datasets)
-            total_words = sum(cnt.values())
-            
-            # 3. Calculate the score for the cluster (laplace rule)
-            score_cluster[lbl] = toxic_words / total_words
-
-        return score_cluster, clusterable_embeddings, embeddings,labels
 
     def _get_guardian_biased_attributes(self, batch:list[Batch], attributes:list[ProtectedAttribute],context:str)-> dict[str,list[BiasMetric.GuardianInteraction]]:
         """
@@ -303,52 +183,30 @@ class Bias(FairForge):
         context: str,
         assistant_id: str,
         batch: list[Batch],
-        language: Optional[str] = "english",
     ):
         """
         Process a batch of interactions to analyze bias.
 
-        This method serves as the main interface for bias analysis, combining various analysis
-        techniques including guardian-based bias detection, clustering, and confidence interval
-        calculations.
+        This method serves as the main interface for bias analysis, using guardian-based
+        bias detection and confidence interval calculations.
 
         Args:
             session_id (str): Unique identifier for the analysis session
             context (str): Context information for bias analysis
             assistant_id (str): Identifier for the AI assistant being analyzed
             batch (list[Batch]): List of batch interactions to analyze
-            language (Optional[str], optional): Language of the responses. Defaults to "english".
         """
-        biases_by_attribute = self._get_guardian_biased_attributes(batch,self.protected_attributes,context)
-        cluster_scores,umap_embeddings,embeddings,labels = self._cluster_profiling(batch,language)
+        biases_by_attribute = self._get_guardian_biased_attributes(batch, self.protected_attributes, context)
 
         self.logger.info(f"Biases by attribute: {biases_by_attribute}")
-        cluster_scores_serializable = {
-          int(k) if isinstance(k, np.integer) else k: float(v) if isinstance(v, np.floating) else v
-          for k, v in cluster_scores.items()
-        }
-        self.logger.info(f"Cluster scores: {cluster_scores_serializable}")
-    
+
         confidence_intervals = self._calculate_confidence_intervals(biases_by_attribute)
 
-        umap_embeddings_serializable = umap_embeddings.tolist() if isinstance(umap_embeddings, np.ndarray) else umap_embeddings
-        embeddings_serializable = embeddings.tolist() if isinstance(embeddings, np.ndarray) else embeddings
-        labels_serializable = labels.tolist() if isinstance(labels, np.ndarray) else labels
-
-        assistant_space = BiasMetric.AssistantSpace(
-            latent_space=umap_embeddings_serializable,
-            embeddings=embeddings_serializable,
-            cluster_labels=labels_serializable
-        )
-
-        
         bias_metric = BiasMetric(
             session_id=session_id,
             assistant_id=assistant_id,
             confidence_intervals=confidence_intervals,
-            guardian_interactions=biases_by_attribute,
-            cluster_profiling=cluster_scores,
-            assistant_space=assistant_space
+            guardian_interactions=biases_by_attribute
         )
         self.metrics.append(bias_metric)
